fix: color and styling
